+++ addon/components/object-list-view.js
@@ -339,14 +339,20 @@ export default FlexberryBaseComponent.extend(
     'showCheckBoxInRow',
     'showDeleteButtonInRow',
     'showEditButtonInRow',
+    'customButtonsInRow',
     'modelProjection',
     function() {
       if (this.get('modelProjection')) {
-        return this.get('showAsteriskInRow') || this.get('showCheckBoxInRow') || this.get('showDeleteButtonInRow') || this.get('showEditButtonInRow');
+        return this.get('showAsteriskInRow') ||
+          this.get('showCheckBoxInRow') ||
+          this.get('showDeleteButtonInRow') ||
+          this.get('showEditButtonInRow') ||
+          !!this.get('customButtonsInRow');
       } else {
         return false;
       }
-    }),
+    }
+  ).readOnly(),
 
   /**
     Flag indicates whether to show dropdown menu with edit menu item, in last column of every row.
@@ -821,6 +827,16 @@ export default FlexberryBaseComponent.extend(
   componentName: '',
 
   actions: {
+    /**
+      Just redirects action up to {{#crossLink "FlexberryObjectlistviewComponent"}}`flexberry-objectlistview`{{/crossLink}} component.
+
+      @method actions.customButtonInRowAction
+      @param {String} actionName The name of action.
+      @param {DS.Model} model Model in row.
+    */
+    customButtonInRowAction(actionName, model) {
+      this.sendAction('customButtonInRowAction', actionName, model);
+    },
 
     /**
       This action is called when user click on row.
@@ -1008,31 +1024,11 @@ export default FlexberryBaseComponent.extend(
       @param {jQuery.Event} e jQuery.Event by click on ckeck all button
     */
     checkAll(e) {
-      let contentWithKeys = this.get('contentWithKeys');
-
       let checked = !this.get('allSelect');
-      Ember.set(this, 'allSelect', checked);
-
-      for (let i = 0; i < contentWithKeys.length; i++) {
-        let recordWithKey = contentWithKeys[i];
-        let selectedRow = this._getRowByKey(recordWithKey.key);
-
-        if (checked) {
-          if (!selectedRow.hasClass('active')) {
-            selectedRow.addClass('active');
-          }
-        } else {
-          if (selectedRow.hasClass('active')) {
-            selectedRow.removeClass('active');
-          }
-        }
-
-        recordWithKey.set('selected', checked);
-        recordWithKey.set('rowConfig.canBeSelected', !checked);
-      }
 
       let componentName = this.get('componentName');
-      this.get('objectlistviewEventsService').updateSelectAllTrigger(componentName, checked);
+      this.get('objectlistviewEventsService').updateSelectAllTrigger(componentName, checked, true);
+      this.selectedRowsChanged();
     },
 
     /**
@@ -1101,10 +1097,12 @@ export default FlexberryBaseComponent.extend(
 
     this.get('objectlistviewEventsService').on('olvAddRow', this, this._addRow);
     this.get('objectlistviewEventsService').on('olvDeleteRows', this, this._deleteRows);
+    this.get('objectlistviewEventsService').on('olvDeleteAllRows', this, this._deleteAllRows);
     this.get('objectlistviewEventsService').on('filterByAnyMatch', this, this._filterByAnyMatch);
     this.get('objectlistviewEventsService').on('refreshList', this, this._refreshList);
     this.get('objectlistviewEventsService').on('geSortApply', this, this._setContent);
     this.get('objectlistviewEventsService').on('updateWidth', this, this.setColumnWidths);
+    this.get('objectlistviewEventsService').on('updateSelectAll', this, this._selectAll);
   },
 
   /**
@@ -1262,10 +1260,12 @@ export default FlexberryBaseComponent.extend(
 
     this.get('objectlistviewEventsService').off('olvAddRow', this, this._addRow);
     this.get('objectlistviewEventsService').off('olvDeleteRows', this, this._deleteRows);
+    this.get('objectlistviewEventsService').off('olvDeleteAllRows', this, this._deleteAllRows);
     this.get('objectlistviewEventsService').off('filterByAnyMatch', this, this._filterByAnyMatch);
     this.get('objectlistviewEventsService').off('refreshList', this, this._refreshList);
     this.get('objectlistviewEventsService').off('geSortApply', this, this._setContent);
     this.get('objectlistviewEventsService').off('updateWidth', this, this.setColumnWidths);
+    this.get('objectlistviewEventsService').off('updateSelectAll', this, this._selectAll);
 
     this.get('objectlistviewEventsService').clearSelectedRecords(this.get('componentName'));
 
@@ -2060,6 +2060,92 @@ export default FlexberryBaseComponent.extend(
   },
 
   /**
+    Handler for "delete all rows on all pages" event in objectlistview.
+
+    @method _deleteRows
+
+    @param {String} componentName The name of objectlistview component
+    @param {Object} filterQuery Filter applying before delete all records on all pages
+  */
+  _deleteAllRows(componentName, filterQuery) {
+    if (componentName === this.get('componentName')) {
+      let currentController = this.get('currentController');
+      currentController.onDeleteActionStarted();
+      let beforeDeleteAllRecords = this.get('beforeDeleteAllRecords');
+      let possiblePromise = null;
+      let modelName = this.get('modelName');
+      let data = {
+        cancel: false,
+        filterQuery: filterQuery
+      };
+
+      if (beforeDeleteAllRecords) {
+        Ember.assert('beforeDeleteAllRecords must be a function', typeof beforeDeleteAllRecords === 'function');
+
+        possiblePromise = beforeDeleteAllRecords(modelName, data);
+
+        if ((!possiblePromise || !(possiblePromise instanceof Ember.RSVP.Promise)) && data.cancel) {
+          return;
+        }
+      }
+
+      if (possiblePromise || (possiblePromise instanceof Ember.RSVP.Promise)) {
+        possiblePromise.then(() => {
+          if (!data.cancel) {
+            this._actualDeleteAllRecords(componentName, modelName, data.filterQuery);
+          }
+        });
+      } else {
+        this._actualDeleteAllRecords(componentName, modelName, data.filterQuery);
+      }
+    }
+  },
+
+  /**
+    Actually delete the all records on all pages.
+
+    @method _actualDeleteAllRecords
+    @private
+
+    @param {String} componentName The name of objectlistview component
+    @param {String} modelName Model name that defines type of records to delete
+    @param {Object} filterQuery Filter applying before delete all records
+  */
+  _actualDeleteAllRecords(componentName, modelName, filterQuery) {
+    let currentController = this.get('currentController');
+    this.get('objectlistviewEventsService').setLoadingState('loading');
+    let promise = this.get('store').deleteAllRecords(modelName, filterQuery);
+
+    promise.then((data)=> {
+      if (data.deletedCount > -1) {
+        this.get('objectlistviewEventsService').setLoadingState('success');
+        this.get('objectlistviewEventsService').rowsDeletedTrigger(componentName, data.deletedCount, true);
+        currentController.onDeleteActionFulfilled();
+        this.get('objectlistviewEventsService').refreshListTrigger(componentName);
+      } else {
+        this.get('objectlistviewEventsService').setLoadingState('error');
+        let errorData = {
+          message: data.message
+        };
+
+        currentController.onDeleteActionRejected(errorData);
+        currentController.send('handleError', errorData);
+      }
+    }).catch((errorData) => {
+      this.get('objectlistviewEventsService').setLoadingState('error');
+      if (!Ember.isNone(errorData.status) && errorData.status === 0 && !Ember.isNone(errorData.statusText) &&  errorData.statusText === 'error') {
+        // This message will be converted to corresponding localized message.
+        errorData.message = 'Ember Data Request returned a 0 Payload (Empty Content-Type)';
+      }
+
+      currentController.onDeleteActionRejected(errorData);
+      this.get('currentController').send('handleError', errorData);
+    }).finally((data) => {
+      currentController.onDeleteActionAlways(data);
+    });
+  },
+
+  /**
     Handler for "delete selected rows" event in objectlistview.
 
     @method _deleteRows
@@ -2078,7 +2164,7 @@ export default FlexberryBaseComponent.extend(
       }, this);
 
       selectedRecords.clear();
-      this.get('objectlistviewEventsService').rowsDeletedTrigger(componentName, count);
+      this.get('objectlistviewEventsService').rowsDeletedTrigger(componentName, count, immediately);
     }
   },
 
@@ -2351,4 +2437,41 @@ export default FlexberryBaseComponent.extend(
     }
   },
 
+  /**
+    Select/Unselect all records on all pages.
+
+    @method _selectAll
+    @private
+  */
+  _selectAll(componentName, selectAllParameter, skipConfugureRows) {
+    if (componentName === this.componentName)
+    {
+      this.set('allSelect', selectAllParameter);
+
+      let contentWithKeys = this.get('contentWithKeys');
+      let selectedRecords = this.get('selectedRecords');
+      for (let i = 0; i < contentWithKeys.length; i++) {
+        let recordWithKey = contentWithKeys[i];
+        let selectedRow = this._getRowByKey(recordWithKey.key);
+
+        if (selectAllParameter) {
+          if (!selectedRow.hasClass('active')) {
+            selectedRow.addClass('active');
+          }
+        } else {
+          if (selectedRow.hasClass('active')) {
+            selectedRow.removeClass('active');
+          }
+        }
+
+        selectedRecords.removeObject(recordWithKey.data);
+        recordWithKey.set('selected', selectAllParameter);
+        recordWithKey.set('rowConfig.canBeSelected', !selectAllParameter);
+      }
+
+      if (!skipConfugureRows) {
+        this.selectedRowsChanged();
+      }
+    }
+  },
 });