+++ tests/dummy/app/controllers/components-examples/flexberry-lookup/lookup-in-modal.js
@@ -2,30 +2,46 @@ import Ember from 'ember';
 import EditFormController from 'ember-flexberry/controllers/edit-form';
 
 export default EditFormController.extend({
+  /**
+    Defaul style of modal context.
+
+    @property readonly
+    @type String
+    @default #example
+  */
+  _style:'#example',
 
   actions: {
-    modalWindow() {
+    modalWindow(style) {
+      if (!Ember.isNone(style)) {
+        this.set('_style', style);
+      }
+
       let repeatWindow = Ember.$('#repeat-window').modal({
         closable: false,
         autofocus: false,
         detachable: false,
         allowMultiple: true,
-        context: '#example',
+        context: this.get('_style'),
       });
       this.set('repeatWindow', repeatWindow);
-      this.get('repeatWindow').modal('refresh').modal('show');
+      this.get('repeatWindow').modal('show').modal('refresh');
     },
 
-    modalWindowDouble() {
+    modalWindowDouble(style) {
+      if (!Ember.isNone(style)) {
+        this.set('_style', style);
+      }
+
       let repeatWindow = Ember.$('#repeat-window-double').modal({
         closable: false,
         autofocus: false,
         detachable: false,
         allowMultiple: true,
-        context: '#example',
+        context: this.get('_style'),
       });
       this.set('repeatWindowdouble', repeatWindow);
-      this.get('repeatWindowdouble').modal('refresh').modal('show');
+      this.get('repeatWindowdouble').modal('show').modal('refresh');
     },
 
     logOut() {
@@ -35,5 +51,6 @@ export default EditFormController.extend({
     logOutDouble() {
       this.get('repeatWindowdouble').modal('hide');
     }
+
   }
 });