+++ tests/dummy/app/controllers/components-examples/flexberry-objectlistview/selected-rows.js
@@ -11,6 +11,15 @@ export default ListFormController.extend({
   editFormRoute: 'ember-flexberry-dummy-suggestion-type-edit',
 
   /**
+    Name of compoent.
+
+    @property componentName
+    @type String
+    @default 'suggestionTypeLocalizedTypesObjectListView'
+   */
+  componentName: 'suggestionTypeLocalizedTypesObjectListView',
+
+  /**
     Count selected rows.
 
     @property countSelectedRows
@@ -21,7 +30,27 @@ export default ListFormController.extend({
 
   actions: {
     configurateSelectedRows(selectedRecords) {
-      this.set('countSelectedRows', selectedRecords.length);
+      if (this.get('allSelect')) {
+        this.set('countSelectedRows', this.model.meta.count);
+      } else {
+        this.set('countSelectedRows', selectedRecords.length);
+      }
     }
-  }
+  },
+
+  _selectAll(componentName, selectAllParameter, skipConfugureRows) {
+    if (componentName === this.get('componentName')) {
+      this.set('allSelect', selectAllParameter);
+    }
+  },
+
+  init: function() {
+    this.get('objectlistviewEventsService').on('updateSelectAll', this, this._selectAll);
+    this._super(...arguments);
+  },
+
+  willDestroyElement() {
+    this.get('objectlistviewEventsService').off('updateSelectAll', this, this._selectAll);
+    this._super(...arguments);
+  },
 });